base: Introduce force background freezer
This will freeze tasks with restricted battery usage when moving them
to background and unfreeze them when moving to foreground. It's an
unsafe freeze compared to the built-in cached app optimizer, which
means victims may miss binder calls or even lead to crashes.

But whatever, it is a real freezer, not a deal of "freeze or die".
The latter one is what the built-in cached app optimizer is doing.
Also, it responds much quicker than the built-in one and it freeze
all backgrounds rather than cached apps only.

Change-Id: I4caedca0aa234f2a402eb389c46e2142b5729772
Signed-off-by: LibXZR <i@xzr.moe>
Signed-off-by: Hưng Phan <phandinhhungvp2001@gmail.com>
Signed-off-by: drkphnx <dark.phnx12@gmail.com>